Move unused prototype utils to utils.deprecated
