+++ virtool_cli/ref/legacy.py
@@ -10,7 +10,7 @@ import click
 from virtool_cli.utils.id_generator import (
     generate_random_alphanumeric,
 )
-from virtool_cli.utils.reference import generate_otu_dirname, is_v1
+from virtool_cli.utils.deprecated.reference import generate_otu_dirname, is_v1
 
 
 @dataclass
