+++ virtool_cli/utils/deprecated/storage.py
@@ -4,9 +4,9 @@ from pathlib import Path
 import aiofiles
 import structlog
 
-from virtool_cli.utils.format import format_isolate
+from virtool_cli.utils.deprecated.format import format_isolate
 from virtool_cli.utils.id_generator import generate_unique_ids
-from virtool_cli.utils.reference import get_isolate_paths, get_sequence_paths
+from virtool_cli.utils.deprecated.reference import get_isolate_paths, get_sequence_paths
 
 
 async def write_records(
